% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dir_fns.R
\docType{data}
\name{dirs}
\alias{dirs}
\title{directory path generation convenience functions}
\format{
An object of class \code{list} of length 13.
}
\usage{
dirs$data()
}
\arguments{
\item{...}{\code{(chr)} directory paths}

\item{mkpath}{\code{lgl} Whether to return a path regardless of whether the file/dir exists or not}

\item{ext}{\code{(chr)} file extension}

\item{mustWork}{\code{lgl} If `TRUE`, an error is given if there are no matching files.}
}
\description{
directory path generation convenience functions
}
\examples{
dirs$data("mydata", ext = "csv")
}
\keyword{datasets}
