% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_dom_elements.R
\name{mc_bg_header}
\alias{mc_bg_header}
\title{Create a Mailchimp header with image background section}
\usage{
mc_bg_header(
  bg_src = "https://www.patapa.dhamma.org/wp-content/uploads/2022/12/aosm-2022.jpeg",
  text = "Annual Old Student Meeting"
)
}
\arguments{
\item{bg_src}{\code{chr} URL of background image to use}

\item{text}{\code{chr} Text of the header}
}
\value{
\code{shiny.tag}
}
\description{
Create a Mailchimp header with image background section
}
\examples{
mc_bg_header()
}
