% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_dom_elements.R
\name{mc_text_content}
\alias{mc_text_content}
\title{Create a Mailchimp text content blcok}
\usage{
mc_text_content(content)
}
\arguments{
\item{content}{\code{shiny.tag.list}}
}
\value{
\code{shiny.tag}
}
\description{
Create a Mailchimp text content blcok
}
\examples{
mc_text_content(tags$p("My Content"))
}
