% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_section.R
\name{section_header}
\alias{section_header}
\title{Create a Mailchimp header section}
\usage{
section_header(el, url = NULL, text = NULL)
}
\arguments{
\item{content}{\code{xml_node}}
}
\value{
\code{xml_doc}
}
\description{
Create a Mailchimp header section
}
