% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_section.R
\name{section_images}
\alias{section_images}
\title{Create a Mailchimp image section}
\usage{
section_images(el, src = NULL)
}
\arguments{
\item{content}{\code{xml_node}}
}
\value{
\code{xml_doc}
}
\description{
Create a Mailchimp image section
}
