% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_section.R
\name{section_text}
\alias{section_text}
\title{Create a Mailchimp text section}
\usage{
section_text(content, boxed = FALSE)
}
\arguments{
\item{content}{\code{xml_node}}
}
\value{
\code{xml_doc}
}
\description{
Create a Mailchimp text section
}
